[STM32][explorer] 增加ROMFS根文件系统
