+++ bsp/stm32/stm32f407-atk-explorer/board/Kconfig
@@ -71,18 +71,26 @@ menu "Onboard Peripheral Drivers"
         select BSP_USING_I2C1
         select PKG_USING_MPU6XXX
 
-    config BSP_USING_SDCARD
-        bool "Enable SDCARD (sdio)"
-        select BSP_USING_SDIO
+    menuconfig BSP_USING_FS
+        bool "Enable filesystem"
         select RT_USING_DFS
-        select RT_USING_DFS_ELMFAT
+        select RT_USING_DFS_ROMFS
         default n
 
-    config SDIO_MAX_FREQ
-        int "sdio max freq"
-        range 0 24000000
-        depends on BSP_USING_SDCARD
-        default 1000000
+        if BSP_USING_FS
+            config BSP_USING_SDCARD
+                bool "Enable SDCARD (FATFS)"
+                select BSP_USING_SDIO
+                select RT_USING_DFS
+                select RT_USING_DFS_ELMFAT
+                default n
+
+            config SDIO_MAX_FREQ
+                int "sdio max freq"
+                range 0 24000000
+                depends on BSP_USING_SDCARD
+                default 1000000
+        endif
 
 endmenu
