+++ bsp/stm32/stm32f407-atk-explorer/board/SConscript
@@ -18,8 +18,8 @@ if GetDepend(['BSP_USING_ETH']):
 if GetDepend(['BSP_USING_SPI_FLASH']):
     src += Glob('ports/spi_flash_init.c')
 
-if GetDepend(['BSP_USING_SDCARD']):
-    src += Glob('ports/drv_sdcard.c')
+if GetDepend(['BSP_USING_FS']):
+    src += Glob('ports/drv_filesystem.c')
 
 if GetDepend(['BSP_USING_SRAM']):
     src += Glob('ports/drv_sram.c')
